requirements.txt
# Web Framework
fastapi>=0.111.0
uvicorn>=0.29.0

# Veritabanı
sqlalchemy>=2.0.0

# Veri Validasyonu
pydantic>=2.0.0

# Form & Dosya Yükleme
python-multipart>=0.0.9

# Zamanlayıcı
apscheduler>=3.10.0

# PDF Üretimi
reportlab>=4.0.0

# Ortam Değişkenleri
python-dotenv>=1.0.0

# Test
pytest>=8.0.0
httpx>=0.27.0
